backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
MaiConverter
python-multipart
python-dotenv
Pillow
tqdm
